Fix C23

+++ C23 - Spring Boot, Dao, Patron MVC/C23S/src/main/java/com/dh/clase23/integral/dao/DomicilioDAOH2.java
@@ -41,7 +41,7 @@ public class DomicilioDAOH2 implements IDao<Domicilio>{
     }
 
     @Override
-    public Domicilio buscar(int id) {
+    public Domicilio buscar(Long id) {
 
         Domicilio domicilio = null;
         Connection connection = null;
