+++ C23 - Spring Boot, Dao, Patron MVC/C23S/src/main/java/com/dh/clase23/integral/dao/IDao.java
@@ -6,7 +6,7 @@ public interface IDao <T> {
 
     void guardar(T t);
 
-    T buscar(int id);
+    T buscar(Long id);
 
     List<T> listarElementos();
